moved classes related to the Trading system into 'trading' package
