+++ app/src/test/java/com/internetEnemies/combatCritters/LogicUnitTests/TradesHandlerTest.java
@@ -12,8 +12,8 @@ package com.internetEnemies.combatCritters.LogicUnitTests;
 import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
-import com.internetEnemies.combatCritters.Logic.ITradesHandler;
-import com.internetEnemies.combatCritters.Logic.TradesHandler;
+import com.internetEnemies.combatCritters.Logic.trading.ITradesHandler;
+import com.internetEnemies.combatCritters.Logic.trading.TradesHandler;
 import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.data.CardInventoryStub;
 import com.internetEnemies.combatCritters.data.CurrencyInventoryStub;
